utilisation_uptime_dashboard/requirements.txt
azure-storage-blob
faicons==0.2.2
geopandas==1.0.1
numpy==2.2.1
pandas==2.2.3
polars==1.12.0
plotly==5.24.1
pyarrow==19.0.0
python-dotenv==1.0.1
pytz==2024.2
shiny==1.2.1
shinywidgets==0.3.2
